Reorganized pytests
