+++ applications_input_files/test_input.punkt
@@ -1,5 +1,3 @@
 main {
-    var x = 0.
-    x = 1 + 3 + 5.
-    print x.
+    x = 5.
 }
