+++ parse_node/parse_nodes/expression_statement_node.h
@@ -9,7 +9,7 @@ class ExpressionStatementNode : public ParseNode {
   ExpressionStatementNode()
       : ParseNode(ParseNodeType::EXPRESSION_STATEMENT_NODE, nullptr) {}
 
-  ParseNode *GetExpression() const { return GetChild(0); }
+  ParseNode *GetExpressionNode() const { return GetChild(0); }
 
   virtual std::string ToString() const override {
     return "EXPRESSION STATEMENT NODE";
